Move deprecations handler to Partners
