+++ packages/cli-kit/src/public/node/api/partners.ts
@@ -1,6 +1,7 @@
 import {partnersFqdn} from '../context/fqdn.js'
-import {graphqlRequest, handleDeprecations, GraphQLVariables} from '../../../private/node/api/graphql.js'
+import {graphqlRequest, GraphQLResponse, GraphQLVariables} from '../../../private/node/api/graphql.js'
 import {gql} from 'graphql-request'
+import {setNextDeprecationDate} from '../../../private/node/conf-store.js'
 
 /**
  * Executes a GraphQL query against the Partners API.
@@ -14,7 +15,8 @@ export async function partnersRequest<T>(query: string, token: string, variables
   const api = 'Partners'
   const fqdn = await partnersFqdn()
   const url = `https://${fqdn}/api/cli/graphql`
-  return graphqlRequest(query, api, url, token, variables)
+  const handleErrors = true
+  return graphqlRequest(query, api, url, token, variables, handleErrors, {deprecationsHandler})
 }
 
 interface ProxyResponse {
@@ -46,7 +48,7 @@ export async function functionProxyRequest<T>(
   const proxyQuery = ScriptServiceProxyQuery
   const proxyResponse: ProxyResponse = await partnersRequest(proxyQuery, token, proxyVariables)
   const scriptServiceResponse = JSON.parse(proxyResponse.scriptServiceProxy)
-  handleDeprecations(scriptServiceResponse)
+  deprecationsHandler(scriptServiceResponse)
 
   return scriptServiceResponse as T
 }
@@ -56,3 +58,19 @@ const ScriptServiceProxyQuery = gql`
     scriptServiceProxy(apiKey: $api_key, query: $query, variables: $variables)
   }
 `
+
+interface Deprecation {
+  supportedUntilDate: string
+}
+
+interface WithDeprecations {
+  deprecations: Deprecation[]
+}
+
+function deprecationsHandler<T>(response: GraphQLResponse<T>) {
+  const deprecations = (response.extensions as WithDeprecations)?.deprecations
+  if (deprecations) {
+    const deprecationDates = deprecations.map(({supportedUntilDate}) => new Date(supportedUntilDate))
+    setNextDeprecationDate(deprecationDates)
+  }
+}
